src/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
nltk>=3.8.1
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0